requirements.txt
fastapi
uvicorn
psycopg2-binary
python-dotenv
pytest
asyncio_mode = auto